feature: viewModel, model 책임 분리
