+++ src/viewModel/ItemObserver.ts
@@ -1,0 +1,6 @@
+import Observable from './Observable';
+import { itemStore } from '../model/ItemStore';
+
+const itemObserver = Observable(itemStore);
+
+export { itemObserver };
